+++ node/src/Worker.ts
@@ -83,6 +83,7 @@ export class WorkerImpl<WorkerAppData extends AppData = AppData>
 		workerBin,
 		libwebrtcFieldTrials,
 		disableLiburing,
+		useBuiltInSctpStack,
 		appData,
 	}: WorkerSettings<WorkerAppData>) {
 		super();
@@ -137,7 +138,13 @@ export class WorkerImpl<WorkerAppData extends AppData = AppData>
 		}
 
 		if (disableLiburing) {
-			spawnArgs.push(`--disableLiburing=true`);
+			spawnArgs.push('--disableLiburing=true');
+		}
+
+		if (useBuiltInSctpStack) {
+			spawnArgs.push('--useBuiltInSctpStack=true');
+		} else {
+			spawnArgs.push('--useBuiltInSctpStack=false');
 		}
 
 		logger.debug(`spawning worker process: ${spawnBin} ${spawnArgs.join(' ')}`);