+++ node/src/index.ts
@@ -101,7 +101,8 @@ export async function createWorker<WorkerAppData extends AppData = AppData>({
 	// eslint-disable-next-line no-shadow
 	workerBin,
 	libwebrtcFieldTrials,
-	disableLiburing,
+	disableLiburing = false,
+	useBuiltInSctpStack = false,
 	appData,
 }: WorkerSettings<WorkerAppData> = {}): Promise<Worker<WorkerAppData>> {
 	logger.debug('createWorker()');
@@ -120,6 +121,7 @@ export async function createWorker<WorkerAppData extends AppData = AppData>({
 		workerBin,
 		libwebrtcFieldTrials,
 		disableLiburing,
+		useBuiltInSctpStack,
 		appData,
 	});