+++ worker/src/RTC/DataConsumer.cpp
@@ -5,9 +5,6 @@
 #include "DepLibUV.hpp"
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
-#ifndef MS_SCTP_STACK
-#include "RTC/SctpAssociation.hpp"
-#endif
 
 namespace RTC
 {
@@ -17,18 +14,12 @@ namespace RTC
 	  RTC::Shared* shared,
 	  const std::string& id,
 	  const std::string& dataProducerId,
-#ifndef MS_SCTP_STACK
-	  RTC::SctpAssociation* sctpAssociation,
-#endif
 	  RTC::DataConsumer::Listener* listener,
 	  const FBS::Transport::ConsumeDataRequest* data,
 	  size_t maxMessageSize)
 	  : id(id),
 	    dataProducerId(dataProducerId),
 	    shared(shared),
-#ifndef MS_SCTP_STACK
-	    sctpAssociation(sctpAssociation),
-#endif
 	    listener(listener),
 	    maxMessageSize(maxMessageSize)
 	{
@@ -222,20 +213,14 @@ namespace RTC
 					MS_THROW_TYPE_ERROR("invalid DataConsumer type");
 				}
 
-#ifdef MS_SCTP_STACK
-				// TODO: SCTP
-#else
-				if (!this->sctpAssociation)
-				{
-					MS_THROW_ERROR("no SCTP association present");
-				}
+				uint32_t bufferedAmount{ 0 };
 
-				// Create status response.
+				this->listener->OnDataConsumerNeedBufferedAmount(this, bufferedAmount);
+
 				auto responseOffset = FBS::DataConsumer::CreateGetBufferedAmountResponse(
-				  request->GetBufferBuilder(), this->sctpAssociation->GetSctpBufferedAmount());
+				  request->GetBufferBuilder(), bufferedAmount);
 
 				request->Accept(FBS::Response::Body::DataConsumer_GetBufferedAmountResponse, responseOffset);
-#endif
 
 				break;
 			}
@@ -284,15 +269,6 @@ namespace RTC
 				{
 					MS_THROW_TYPE_ERROR("invalid DataConsumer type");
 				}
-
-#ifdef MS_SCTP_STACK
-				// TODO: SCTP
-#else
-				if (!this->sctpAssociation)
-				{
-					MS_THROW_ERROR("no SCTP association present");
-				}
-#endif
 
 				const auto* body    = request->data->body_as<FBS::DataConsumer::SendRequest>();
 				const uint8_t* data = body->data()->Data();
@@ -480,7 +456,7 @@ namespace RTC
 		MS_DEBUG_DEV("SctpAssociation closed [dataConsumerId:%s]", this->id.c_str());
 	}
 
-	void DataConsumer::SctpAssociationBufferedAmount(uint32_t bufferedAmount)
+	void DataConsumer::SetSctpAssociationBufferedAmount(uint32_t bufferedAmount)
 	{
 		MS_TRACE();