+++ worker/src/RTC/Transport.cpp
@@ -4,6 +4,7 @@
 #include "RTC/Transport.hpp"
 #include "Logger.hpp"
 #include "MediaSoupErrors.hpp"
+#include "Settings.hpp"
 #include "Utils.hpp"
 #ifdef MS_LIBURING_SUPPORTED
 #include "DepLibUring.hpp"
@@ -19,10 +20,8 @@
 #include "RTC/RTCP/FeedbackRtpTransport.hpp"
 #include "RTC/RTCP/XrDelaySinceLastRr.hpp"
 #include "RTC/RtpDictionaries.hpp"
-#ifdef MS_SCTP_STACK
 #include "RTC/SCTP/association/Association.hpp"
 #include "RTC/SCTP/public/SctpOptions.hpp"
-#endif
 #include "RTC/SimpleConsumer.hpp"
 #include "RTC/SimulcastConsumer.hpp"
 #include "RTC/SvcConsumer.hpp"
@@ -35,8 +34,8 @@
 
 namespace RTC
 {
-	static const size_t DefaultSctpSendBufferSize{ 262144 }; // 2^18.
-	static const size_t MaxSctpSendBufferSize{ 268435456 };  // 2^28.
+	static const size_t DefaultSctpSendBufferSize{ 262144 }; // 2^18 bytes.
+	static const size_t MaxSctpSendBufferSize{ 268435456 };  // 2^28 bytes.
 
 	/* Instance methods. */
 
@@ -111,28 +110,32 @@ namespace RTC
 				sctpSendBufferSize = DefaultSctpSendBufferSize;
 			}
 
-#ifdef MS_SCTP_STACK
-			// TODO: SCTP: Many interesting options missing.
-			const RTC::SCTP::SctpOptions sctpOptions = { .sourcePort         = 5000,
-				                                           .destinationPort    = 5000,
-				                                           .maxOutboundStreams = 65535,
-				                                           .maxInboundStreams =
-				                                             options->numSctpStreams()->mis(),
-				                                           // TODO: SCTP: Sure?
-				                                           .maxSendMessageSize = this->maxMessageSize,
-				                                           .maxSendBufferSize  = sctpSendBufferSize };
+			if (Settings::configuration.useBuiltInSctpStack)
+			{
+				// TODO: SCTP: Many interesting options missing.
+				const RTC::SCTP::SctpOptions sctpOptions = { .sourcePort         = 5000,
+					                                           .destinationPort    = 5000,
+					                                           .maxOutboundStreams = 65535,
+					                                           .maxInboundStreams =
+					                                             options->numSctpStreams()->mis(),
+					                                           // TODO: SCTP: Sure?
+					                                           .maxSendMessageSize = this->maxMessageSize,
+					                                           .maxSendBufferSize  = sctpSendBufferSize };
 
-			this->sctpAssociation = std::make_unique<RTC::SCTP::Association>(sctpOptions, this);
-#else
-			// This may throw.
-			this->sctpAssociation = new RTC::SctpAssociation(
-			  this,
-			  options->numSctpStreams()->os(),
-			  options->numSctpStreams()->mis(),
-			  this->maxMessageSize,
-			  sctpSendBufferSize,
-			  options->isDataChannel());
-#endif
+				this->sctpAssociation = std::make_unique<RTC::SCTP::Association>(sctpOptions, this);
+			}
+			// TODO: Remove once we only use built-in SCTP stack.
+			else
+			{
+				// This may throw.
+				this->oldSctpAssociation = new RTC::SctpAssociation(
+				  this,
+				  options->numSctpStreams()->os(),
+				  options->numSctpStreams()->mis(),
+				  this->maxMessageSize,
+				  sctpSendBufferSize,
+				  options->isDataChannel());
+			}
 		}
 
 		// Create the RTCP timer.
@@ -183,17 +186,23 @@ namespace RTC
 		}
 		this->mapDataConsumers.clear();
 
-#ifdef MS_SCTP_STACK
-		// NOTE: We don't do anything here since the `Destroying()` method has already
-		// been called by the Transport subclass and it closes the SCTP Association.
-		// NOTE: We cannot do it here in the destructor because here we are no longer
-		// the Transport subclass but Transport parent (this is how the destruction
-		// chain works in C++).
-#else
-		// Delete SCTP association.
-		delete this->sctpAssociation;
-		this->sctpAssociation = nullptr;
-#endif
+		if (Settings::configuration.useBuiltInSctpStack)
+		{
+			// NOTE: When using the built-in SCTP stack we don't do anything here since
+			// the `Destroying()` method has already been called by the Transport subclas
+			// and it closed the SCTP Association.
+			// NOTE: We cannot do it here in the destructor because here we are no longer
+			// the Transport subclass but Transport parent (this is how the destruction
+			// chain works in C++).
+		}
+		// TODO: Remove once we only use built-in SCTP stack.
+		else
+		{
+			// Delete SCTP association.
+			// TODO: Remove once we only use built-in SCTP stack.
+			delete this->oldSctpAssociation;
+			this->oldSctpAssociation = nullptr;
+		}
 
 		// Delete the RTCP timer.
 		delete this->rtcpTimer;
@@ -366,12 +375,11 @@ namespace RTC
 		// Add sctpListener.
 		flatbuffers::Offset<FBS::Transport::SctpListener> sctpListener;
 
-		if (this->sctpAssociation)
+		if (Settings::configuration.useBuiltInSctpStack && this->sctpAssociation)
 		{
 			// Add sctpParameters.
 			sctpParameters = this->sctpAssociation->FillBuffer(builder);
 
-#ifdef MS_SCTP_STACK
 			// NOTE: There is never permanent FAILED state.
 			switch (this->sctpAssociation->GetAssociationState())
 			{
@@ -400,8 +408,16 @@ namespace RTC
 					break;
 				}
 			}
-#else
-			switch (this->sctpAssociation->GetState())
+
+			sctpListener = this->sctpListener.FillBuffer(builder);
+		}
+		// TODO: Remove once we only use built-in SCTP stack.
+		else if (!Settings::configuration.useBuiltInSctpStack && this->oldSctpAssociation)
+		{
+			// Add sctpParameters.
+			sctpParameters = this->oldSctpAssociation->FillBuffer(builder);
+
+			switch (this->oldSctpAssociation->GetState())
 			{
 				case RTC::SctpAssociation::SctpState::NEW:
 				{
@@ -433,7 +449,7 @@ namespace RTC
 					break;
 				}
 			}
-#endif
+
 			sctpListener = this->sctpListener.FillBuffer(builder);
 		}
 
@@ -463,8 +479,9 @@ namespace RTC
 		  rtpListenerOffset,
 		  this->maxMessageSize,
 		  sctpParameters,
-		  this->sctpAssociation ? flatbuffers::Optional<FBS::SctpAssociation::SctpState>(sctpState)
-		                        : flatbuffers::nullopt,
+		  (this->sctpAssociation || this->oldSctpAssociation)
+		    ? flatbuffers::Optional<FBS::SctpAssociation::SctpState>(sctpState)
+		    : flatbuffers::nullopt,
 		  sctpListener,
 		  &traceEventTypes);
 	}
@@ -479,10 +496,9 @@ namespace RTC
 		// Add sctpState.
 		FBS::SctpAssociation::SctpState sctpState{ FBS::SctpAssociation::SctpState::NEW };
 
-		if (this->sctpAssociation)
+		// Add sctpState.
+		if (Settings::configuration.useBuiltInSctpStack && this->sctpAssociation)
 		{
-			// Add sctpState.
-#ifdef MS_SCTP_STACK
 			// NOTE: There is never permanent FAILED state.
 			switch (this->sctpAssociation->GetAssociationState())
 			{
@@ -511,8 +527,11 @@ namespace RTC
 					break;
 				}
 			}
-#else
-			switch (this->sctpAssociation->GetState())
+		}
+		// TODO: Remove once we only use built-in SCTP stack.
+		else if (!Settings::configuration.useBuiltInSctpStack && this->oldSctpAssociation)
+		{
+			switch (this->oldSctpAssociation->GetState())
 			{
 				case RTC::SctpAssociation::SctpState::NEW:
 				{
@@ -544,7 +563,6 @@ namespace RTC
 					break;
 				}
 			}
-#endif
 		}
 
 		return FBS::Transport::CreateStatsDirect(
@@ -554,8 +572,9 @@ namespace RTC
 		  // timestamp.
 		  nowMs,
 		  // sctpState.
-		  this->sctpAssociation ? flatbuffers::Optional<FBS::SctpAssociation::SctpState>(sctpState)
-		                        : flatbuffers::nullopt,
+		  (this->sctpAssociation || this->oldSctpAssociation)
+		    ? flatbuffers::Optional<FBS::SctpAssociation::SctpState>(sctpState)
+		    : flatbuffers::nullopt,
 		  // bytesReceived.
 		  this->recvTransmission.GetBytes(),
 		  // recvBitrate.
@@ -1137,7 +1156,10 @@ namespace RTC
 			case Channel::ChannelRequest::Method::TRANSPORT_PRODUCE_DATA:
 			{
 				// Early check. The Transport must support SCTP or be direct.
-				if (!this->sctpAssociation && !this->direct)
+				if (
+				  ((Settings::configuration.useBuiltInSctpStack && !this->sctpAssociation) ||
+				   (!Settings::configuration.useBuiltInSctpStack && !this->oldSctpAssociation)) &&
+				  !this->direct)
 				{
 					MS_THROW_ERROR("SCTP not enabled and not a direct Transport");
 				}
@@ -1158,7 +1180,9 @@ namespace RTC
 				{
 					case RTC::DataProducer::Type::SCTP:
 					{
-						if (!this->sctpAssociation)
+						if (
+						  (Settings::configuration.useBuiltInSctpStack && !this->sctpAssociation) ||
+						  (!Settings::configuration.useBuiltInSctpStack && !this->oldSctpAssociation))
 						{
 							delete dataProducer;
 
@@ -1231,11 +1255,15 @@ namespace RTC
 				if (dataProducer->GetType() == RTC::DataProducer::Type::SCTP)
 				{
 					// Tell to the SCTP association.
-#ifdef MS_SCTP_STACK
-					this->sctpAssociation->MayConnect();
-#else
-					this->sctpAssociation->HandleDataProducer(dataProducer);
-#endif
+					if (Settings::configuration.useBuiltInSctpStack)
+					{
+						this->sctpAssociation->MayConnect();
+					}
+					// TODO: Remove once we only use built-in SCTP stack.
+					else
+					{
+						this->oldSctpAssociation->HandleDataProducer(dataProducer);
+					}
 				}
 
 				break;
@@ -1244,7 +1272,10 @@ namespace RTC
 			case Channel::ChannelRequest::Method::TRANSPORT_CONSUME_DATA:
 			{
 				// Early check. The Transport must support SCTP or be direct.
-				if (!this->sctpAssociation && !this->direct)
+				if (
+				  ((Settings::configuration.useBuiltInSctpStack && !this->sctpAssociation) ||
+				   (!Settings::configuration.useBuiltInSctpStack && !this->oldSctpAssociation)) &&
+				  !this->direct)
 				{
 					MS_THROW_ERROR("SCTP not enabled and not a direct Transport");
 				}
@@ -1259,22 +1290,16 @@ namespace RTC
 
 				// This may throw.
 				auto* dataConsumer = new RTC::DataConsumer(
-				  this->shared,
-				  dataConsumerId,
-				  dataProducerId,
-#ifndef MS_SCTP_STACK
-				  this->sctpAssociation,
-#endif
-				  this,
-				  body,
-				  this->maxMessageSize);
+				  this->shared, dataConsumerId, dataProducerId, this, body, this->maxMessageSize);
 
 				// Verify the type of the DataConsumer.
 				switch (dataConsumer->GetType())
 				{
 					case RTC::DataConsumer::Type::SCTP:
 					{
-						if (!this->sctpAssociation)
+						if (
+						  (Settings::configuration.useBuiltInSctpStack && !this->sctpAssociation) ||
+						  (!Settings::configuration.useBuiltInSctpStack && !this->oldSctpAssociation))
 						{
 							delete dataConsumer;
 
@@ -1333,21 +1358,29 @@ namespace RTC
 
 				if (dataConsumer->GetType() == RTC::DataConsumer::Type::SCTP)
 				{
-#ifdef MS_SCTP_STACK
-					if (this->sctpAssociation->GetAssociationState() == RTC::SCTP::Types::AssociationState::CONNECTED)
-#else
-					if (this->sctpAssociation->GetState() == RTC::SctpAssociation::SctpState::CONNECTED)
-#endif
+					if (Settings::configuration.useBuiltInSctpStack)
 					{
-						dataConsumer->SctpAssociationConnected();
-					}
+						if (this->sctpAssociation->GetAssociationState() == RTC::SCTP::Types::AssociationState::CONNECTED)
+						{
+							// Tell to the DataConsumer.
+							dataConsumer->SctpAssociationConnected();
+						}
 
-					// Tell to the SCTP association.
-#ifdef MS_SCTP_STACK
-					this->sctpAssociation->MayConnect();
-#else
-					this->sctpAssociation->HandleDataConsumer(dataConsumer);
-#endif
+						// Tell to the SCTP association.
+						this->sctpAssociation->MayConnect();
+					}
+					// TODO: Remove once we only use built-in SCTP stack.
+					else
+					{
+						if (this->oldSctpAssociation->GetState() == RTC::SctpAssociation::SctpState::CONNECTED)
+						{
+							// Tell to the DataConsumer.
+							dataConsumer->SctpAssociationConnected();
+						}
+
+						// Tell to the SCTP association.
+						this->oldSctpAssociation->HandleDataConsumer(dataConsumer);
+					}
 				}
 
 				break;
@@ -1474,7 +1507,9 @@ namespace RTC
 
 			case Channel::ChannelRequest::Method::TRANSPORT_CLOSE_DATAPRODUCER:
 			{
-				if (!this->sctpAssociation)
+				if (
+				  (Settings::configuration.useBuiltInSctpStack && !this->sctpAssociation) ||
+				  (!Settings::configuration.useBuiltInSctpStack && !this->oldSctpAssociation))
 				{
 					MS_WARN_TAG(sctp, "cannot close a DataProducer, no SCTP Association");
 					;
@@ -1505,12 +1540,16 @@ namespace RTC
 
 				if (dataProducer->GetType() == RTC::DataProducer::Type::SCTP)
 				{
-#ifdef MS_SCTP_STACK
-					// TODO: SCTP
-#else
-					// Tell the SctpAssociation so it can reset the SCTP stream.
-					this->sctpAssociation->DataProducerClosed(dataProducer);
-#endif
+					if (Settings::configuration.useBuiltInSctpStack)
+					{
+						// TODO: SCTP
+					}
+					// TODO: Remove once we only use built-in SCTP stack.
+					else
+					{
+						// Tell the SctpAssociation so it can reset the SCTP stream.
+						this->oldSctpAssociation->DataProducerClosed(dataProducer);
+					}
 				}
 
 				// Delete it.
@@ -1523,7 +1562,9 @@ namespace RTC
 
 			case Channel::ChannelRequest::Method::TRANSPORT_CLOSE_DATACONSUMER:
 			{
-				if (!this->sctpAssociation)
+				if (
+				  (Settings::configuration.useBuiltInSctpStack && !this->sctpAssociation) ||
+				  (!Settings::configuration.useBuiltInSctpStack && !this->oldSctpAssociation))
 				{
 					MS_WARN_TAG(sctp, "cannot close a DataConsumer, no SCTP Association");
 
@@ -1547,12 +1588,16 @@ namespace RTC
 
 				if (dataConsumer->GetType() == RTC::DataConsumer::Type::SCTP)
 				{
-#ifdef MS_SCTP_STACK
-					// TODO: SCTP
-#else
-					// Tell the SctpAssociation so it can reset the SCTP stream.
-					this->sctpAssociation->DataConsumerClosed(dataConsumer);
-#endif
+					if (Settings::configuration.useBuiltInSctpStack)
+					{
+						// TODO: SCTP
+					}
+					// TODO: Remove once we only use built-in SCTP stack.
+					else
+					{
+						// Tell the SctpAssociation so it can reset the SCTP stream.
+						this->oldSctpAssociation->DataConsumerClosed(dataConsumer);
+					}
 				}
 
 				// Delete it.
@@ -1597,16 +1642,17 @@ namespace RTC
 	{
 		MS_TRACE();
 
-#ifdef MS_SCTP_STACK
-		if (this->sctpAssociation)
+		if (Settings::configuration.useBuiltInSctpStack)
 		{
-			// NOTE: We don't invoke `Shutdown()` but `Close()` in the SCTP Association
-			// because at this point we are closing everything and we won't have any
-			// chance to complete the SCTP SHUTDOWN + SHUTDOWN_ACK + SHUTDOWN_COMPLETE
-			// dance, so we invoke `Close()` which just sends a SCTP ABORT.
-			this->sctpAssociation->Close();
+			if (this->sctpAssociation)
+			{
+				// NOTE: We don't invoke `Shutdown()` but `Close()` in the SCTP Association
+				// because at this point we are closing everything and we won't have any
+				// chance to complete the SCTP SHUTDOWN + SHUTDOWN_ACK + SHUTDOWN_COMPLETE
+				// dance, so we invoke `Close()` which just sends a SCTP ABORT.
+				this->sctpAssociation->Close();
+			}
 		}
-#endif
 
 		this->destroying = true;
 	}
@@ -1632,13 +1678,14 @@ namespace RTC
 		}
 
 		// Tell the SctpAssociation.
-		if (this->sctpAssociation)
+		if (Settings::configuration.useBuiltInSctpStack && this->sctpAssociation)
 		{
-#ifdef MS_SCTP_STACK
 			this->sctpAssociation->MayConnect();
-#else
-			this->sctpAssociation->TransportConnected();
-#endif
+		}
+		// TODO: Remove once we only use built-in SCTP stack.
+		else if (!Settings::configuration.useBuiltInSctpStack && this->oldSctpAssociation)
+		{
+			this->oldSctpAssociation->TransportConnected();
 		}
 
 		// Start the RTCP timer.
@@ -1685,12 +1732,11 @@ namespace RTC
 			dataConsumer->TransportDisconnected();
 		}
 
+		// TODO: Remove once we only use built-in SCTP stack.
 		// Tell the SctpAssociation.
-		if (this->sctpAssociation)
+		if (!Settings::configuration.useBuiltInSctpStack && this->oldSctpAssociation)
 		{
-#ifndef MS_SCTP_STACK
-			this->sctpAssociation->TransportDisconnected();
-#endif
+			this->oldSctpAssociation->TransportDisconnected();
 		}
 
 		// Stop the RTCP timer.
@@ -1817,7 +1863,9 @@ namespace RTC
 	{
 		MS_TRACE();
 
-		if (!this->sctpAssociation)
+		if (
+		  (Settings::configuration.useBuiltInSctpStack && !this->sctpAssociation) ||
+		  (!Settings::configuration.useBuiltInSctpStack && !this->oldSctpAssociation))
 		{
 			MS_DEBUG_TAG(sctp, "ignoring SCTP packet (SCTP not enabled)");
 
@@ -1825,11 +1873,14 @@ namespace RTC
 		}
 
 		// Pass it to the SctpAssociation.
-#ifdef MS_SCTP_STACK
-		this->sctpAssociation->ReceiveSctpData(data, len);
-#else
-		this->sctpAssociation->ProcessSctpData(data, len);
-#endif
+		if (Settings::configuration.useBuiltInSctpStack)
+		{
+			this->sctpAssociation->ReceiveSctpData(data, len);
+		}
+		else
+		{
+			this->oldSctpAssociation->ProcessSctpData(data, len);
+		}
 	}
 
 	void Transport::CheckNoDataProducer(const std::string& dataProducerId) const
@@ -2910,6 +2961,31 @@ namespace RTC
 		SendMessage(dataConsumer, msg, len, ppid, cb);
 	}
 
+	void Transport::OnDataConsumerNeedBufferedAmount(
+	  RTC::DataConsumer* /*dataConsumer*/, uint32_t& bufferedAmount)
+	{
+		MS_TRACE();
+
+		if (Settings::configuration.useBuiltInSctpStack && this->sctpAssociation)
+		{
+			// TODO: SCTP
+			// TODO: Let's see how to obtain `streamId` argument from the DataConsumer.
+			// bufferedAmount = this->sctpAssociation->GetStreamBufferedAmount(streamId);
+		}
+		// TODO: Remove once we only use built-in SCTP stack.
+		else if (!Settings::configuration.useBuiltInSctpStack && this->oldSctpAssociation)
+		{
+			// NOTE: The underlaying SCTP association uses a common send buffer for all
+			// data consumers, hence the value given by this method indicates the data
+			// buffered for all data consumers in the transport.
+			bufferedAmount = this->oldSctpAssociation->GetSctpBufferedAmount();
+		}
+		else
+		{
+			bufferedAmount = 0;
+		}
+	}
+
 	void Transport::OnDataConsumerDataProducerClosed(RTC::DataConsumer* dataConsumer)
 	{
 		MS_TRACE();
@@ -2920,21 +2996,27 @@ namespace RTC
 		// Notify the listener.
 		this->listener->OnTransportDataConsumerDataProducerClosed(this, dataConsumer);
 
-		if (this->sctpAssociation && dataConsumer->GetType() == RTC::DataConsumer::Type::SCTP)
+		if (Settings::configuration.useBuiltInSctpStack)
 		{
-#ifdef MS_SCTP_STACK
-			// TODO: SCTP
-#else
-			// Tell the SctpAssociation so it can reset the SCTP stream.
-			this->sctpAssociation->DataConsumerClosed(dataConsumer);
-#endif
+			if (this->sctpAssociation && dataConsumer->GetType() == RTC::DataConsumer::Type::SCTP)
+			{
+				// TODO: SCTP
+			}
+		}
+		// TODO: Remove once we only use built-in SCTP stack.
+		else
+		{
+			if (this->oldSctpAssociation && dataConsumer->GetType() == RTC::DataConsumer::Type::SCTP)
+			{
+				// Tell the SctpAssociation so it can reset the SCTP stream.
+				this->oldSctpAssociation->DataConsumerClosed(dataConsumer);
+			}
 		}
 
 		// Delete it.
 		delete dataConsumer;
 	}
 
-#ifdef MS_SCTP_STACK
 	bool Transport::OnAssociationSendData(const uint8_t* data, size_t len)
 	{
 		MS_TRACE();
@@ -3133,7 +3215,7 @@ namespace RTC
 	}
 
 	// TODO: SCTP: Add OnAssociationLifecycleMessageXxxxxx() methods.
-#else
+
 	void Transport::OnSctpAssociationConnecting(RTC::SctpAssociation* /*sctpAssociation*/)
 	{
 		MS_TRACE();
@@ -3294,11 +3376,10 @@ namespace RTC
 
 			if (dataConsumer->GetType() == RTC::DataConsumer::Type::SCTP)
 			{
-				dataConsumer->SctpAssociationBufferedAmount(bufferedAmount);
+				dataConsumer->SetSctpAssociationBufferedAmount(bufferedAmount);
 			}
 		}
 	}
-#endif
 
 	void Transport::OnTransportCongestionControlClientBitrates(
 	  RTC::TransportCongestionControlClient* /*tccClient*/,