+++ rust/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # NEXT
 
+- Worker: Add `use_built_in_sctp_stack` setting (defaults to `false`) to enable mediasoup built-in SCTP stack ([PR #XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
+
 # 0.21.0
 
 - `router.pipe_producer_to_router()` and `router.pipe_data_producer_to_router()` can now connect two `Routers` in the same `Worker` if `keep_id` is set to `false` (PR #1604).